Make /etc/pwrstatd.conf writable by root
Occasionally pwrstatd attempts to write out its current config,
notably when performing a test of the UPS. When /etc/pwrstatd.conf
is set using environment.etc.<?>.text, pwrstatd segfaults upon calling
fwrite on a null file pointer.

To work around this issue, we write to the file every time the systemd
service is started. This does mean that between service restarts, the
config can be modified using the pwrstat client, but for now that is
acceptable.

Not sure if the references to pkgs.coreutils are necessary when using
the script option for systemd. Placing the same script in ExecStart did
not appear to work.