Revert creating pwrstatd.conf on powerpanel.service starting
Using environment.etc."pwrstatd.conf".mode, the file can be copied
instead of symlinked into /etc. This allows pwrstatd to dump its config
without segfaulting, and still maintains the declarative file placement.

In practical terms, this means that the config file will be refreshed
from the module configuration whenever nixos-rebuild switch is run.
Previously it would be refreshed whenever the service started. Not sure
which is better, but this solution feels cleaner.